refactor: yoga-1.png 목업 이미지를 shared/assets/images로 통합
세 페이지(home/daily-routine/pose-challenge)가 동일한 파일을 각자 assets 폴더에
중복 보관하던 것을 shared/assets/images로 옮겨 하나만 import하도록 정리.
